Split functions used only on client
