+++ client/components/shared/utils.js
@@ -1,0 +1,45 @@
+/* global window */
+import moment from 'moment';
+
+moment.locale('pt-br');
+
+export const STORAGEKEY = 'storedTimes';
+export const STORAGEDAYKEY = 'storedMoment';
+
+export const areTheSameDay = (date1, date2) => (
+	date1.day() === date2.day() &&
+	date1.month() === date2.month() &&
+	date1.year() === date2.year()
+);
+
+export const replacingValueInsideArray = (array, index, newValue) => [
+	...array.slice(0, index),
+	newValue,
+	...array.slice(index + 1)
+];
+
+const _setStorage = (key, data) => {
+	window.localStorage.setItem(key, JSON.stringify(data));
+};
+
+export const setTodayStorage = (key, dayKey, data) => {
+	const today = moment();
+	_setStorage(dayKey, today);
+	_setStorage(key, data);
+};
+
+const _getStorage = key => (
+	JSON.parse(window.localStorage.getItem(key))
+);
+
+export const getTodayStorage = (key, dayKey) => {
+	const dayOnLocal = moment(_getStorage(dayKey));
+	const today = moment();
+	if (dayKey in window.localStorage) {
+		if (areTheSameDay(dayOnLocal, today)) {
+			return _getStorage(key);
+		}
+	}
+	setTodayStorage(key, dayKey, { storedTimes: [{}, {}, {}, {}], sentToday: false });
+	return { storedTimes: [{}, {}, {}, {}], sentToday: false };
+};
