+++ shared/utils.js
@@ -1,11 +1,3 @@
-/* global window */
-import moment from 'moment';
-
-moment.locale('pt-br');
-
-const STORAGEKEY = 'storedTimes';
-const STORAGEDAYKEY = 'storedMoment';
-
 const _isNumber = value => !Number.isNaN(parseInt(value, 10));
 
 const checkValidity = (mode, value) => {
@@ -38,54 +30,9 @@ const buildDateFromTimeString = (timeString) => {
 	return new Date(year, month, day, hours, minutes);
 };
 
-const areTheSameDay = (date1, date2) => (
-	date1.day() === date2.day() &&
-	date1.month() === date2.month() &&
-	date1.year() === date2.year()
-);
-
-const _getStorage = key => (
-	JSON.parse(window.localStorage.getItem(key))
-);
-
-const _setStorage = (key, data) => {
-	window.localStorage.setItem(key, JSON.stringify(data));
-};
-
-const setTodayStorage = (key, dayKey, data) => {
-	const today = moment();
-	_setStorage(dayKey, today);
-	_setStorage(key, data);
-};
-
-const getTodayStorage = (key, dayKey) => {
-	const dayOnLocal = moment(_getStorage(dayKey));
-	const today = moment();
-	if (dayKey in window.localStorage) {
-		if (areTheSameDay(dayOnLocal, today)) {
-			return _getStorage(key);
-		}
-	}
-	setTodayStorage(key, dayKey, [{}, {}, {}, {}]);
-	return [{}, {}, {}, {}];
-};
-
-const replacingValueInsideArray = (array, index, newValue) => [
-	...array.slice(0, index),
-	newValue,
-	...array.slice(index + 1)
-];
-
-
 module.exports = {
 	checkValidity,
 	timeIsValid,
 	getTimeFromDate,
-	buildDateFromTimeString,
-	areTheSameDay,
-	getTodayStorage,
-	setTodayStorage,
-	replacingValueInsideArray,
-	STORAGEKEY,
-	STORAGEDAYKEY
+	buildDateFromTimeString
 };
